+++ packages/jsii-pacmak/test/expected.jsii-calc-base/java/src/main/java/software/amazon/jsii/tests/calculator/base/$Module.java
@@ -14,4 +14,13 @@ public final class $Module extends JsiiModule {
     public List<Class<? extends JsiiModule>> getDependencies() {
         return asList(software.amazon.jsii.tests.calculator.baseofbase.$Module.class);
     }
+
+    @Override
+    protected Class<?> resolveClass(final String fqn) throws ClassNotFoundException {
+        switch (fqn) {
+            case "@scope/jsii-calc-base.Base": return software.amazon.jsii.tests.calculator.base.Base.class;
+            case "@scope/jsii-calc-base.BaseProps": return software.amazon.jsii.tests.calculator.base.BaseProps.class;
+            default: throw new ClassNotFoundException("Unknown JSII type: " + fqn);
+        }
+    }
 }
