+++ packages/jsii-pacmak/test/expected.jsii-calc/dotnet/Amazon.JSII.Tests.Calculator/.jsii
@@ -2791,17 +2791,20 @@
           },
           "name": "stringStyle",
           "type": {
-            "fqn": "jsii-calc.composition.CompositionStringStyle"
+            "fqn": "jsii-calc.composition.CompositeOperation.CompositionStringStyle"
           }
         }
+      ],
+      "subtypes": [
+        "jsii-calc.composition.CompositeOperation.CompositionStringStyle"
       ]
     },
-    "jsii-calc.composition.CompositionStringStyle": {
+    "jsii-calc.composition.CompositeOperation.CompositionStringStyle": {
       "assembly": "jsii-calc",
       "docs": {
         "comment": "Style of .toString() output for CompositeOperation."
       },
-      "fqn": "jsii-calc.composition.CompositionStringStyle",
+      "fqn": "jsii-calc.composition.CompositeOperation.CompositionStringStyle",
       "kind": "enum",
       "members": [
         {
@@ -2812,9 +2815,10 @@
         }
       ],
       "name": "CompositionStringStyle",
-      "namespace": "jsii-calc.composition"
+      "namespace": "jsii-calc.composition.CompositeOperation",
+      "parenttype": "jsii-calc.composition.CompositeOperation"
     }
   },
   "version": "0.6.4",
-  "fingerprint": "ukdkUVnf3jHKB7TV8FhG/tHnl04+DaTZSoIEZxym6Js="
+  "fingerprint": "V1fwQARJzmO/FTBQutFnq1v95it317kCGVv350RLqYo="
 }
