--- packages/jsii-pacmak/test/expected.jsii-calc/java/src/main/java/software/amazon/jsii/tests/calculator/composition/CompositionStringStyle.java
@@ -1,11 +1,0 @@
-package software.amazon.jsii.tests.calculator.composition;
-
-/**
- * Style of .toString() output for CompositeOperation.
- */
-@javax.annotation.Generated(value = "jsii-pacmak")
-@software.amazon.jsii.Jsii(module = software.amazon.jsii.tests.calculator.$Module.class, fqn = "jsii-calc.composition.CompositionStringStyle")
-public enum CompositionStringStyle {
-    Normal,
-    Decorated,
-}
